# apps/api/faulttrace_api/routes/datasets.py
"""
Dataset snapshot REST API endpoints — Prompt 2 (WP8).
"""

from __future__ import annotations

from pathlib import Path
from typing import Any, Optional

from fastapi import APIRouter, HTTPException, Query

from faulttrace_api.config import get_settings

router = APIRouter()


def _get_registry():
    from faulttrace_data.snapshot import SnapshotRegistry
    settings = get_settings()
    registry_path = settings.data_root / "manifests" / "snapshots.jsonl"
    return SnapshotRegistry(registry_path)


@router.get("/datasets", summary="List all ingested dataset snapshots")
async def list_datasets(
    dataset_id: Optional[str] = Query(None, description="Filter by dataset ID"),
    active_only: bool = Query(True, description="Only return active snapshots"),
    page: int = Query(1, ge=1),
    page_size: int = Query(50, ge=1, le=500),
) -> dict[str, Any]:
    try:
        registry = _get_registry()
        snapshots = registry.list_snapshots(dataset_id=dataset_id, active_only=active_only)
        
        total = len(snapshots)
        start = (page - 1) * page_size
        end = start + page_size
        page_snapshots = snapshots[start:end]
        
        return {
            "items": [
                {
                    "snapshot_id": s.snapshot_id,
                    "dataset_id": s.dataset_id,
                    "source_type": s.source_type,
                    "row_count": s.row_count,
                    "accepted_count": s.accepted_count,
                    "rejected_count": s.rejected_count,
                    "duplicate_count": s.duplicate_count,
                    "malformed_count": s.malformed_count,
                    "active": s.active,
                    "created_at": s.created_at,
                    "license_note": s.license_note,
                    "canonical_content_hash": s.canonical_content_hash[:16] if s.canonical_content_hash else "",
                }
                for s in page_snapshots
            ],
            "total": total,
            "page": page,
            "page_size": page_size,
            "has_next": end < total,
            "count": len(page_snapshots), # For backwards compatibility
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@router.get("/datasets/{snapshot_id}", summary="Get a specific snapshot by ID")
async def get_snapshot(snapshot_id: str) -> dict[str, Any]:
    try:
        registry = _get_registry()
        snapshot = registry.inspect(snapshot_id)
        if snapshot is None:
            raise HTTPException(status_code=404, detail=f"Snapshot '{snapshot_id}' not found")
        return snapshot.model_dump()
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@router.get("/datasets/{snapshot_id}/validate", summary="Validate a snapshot's integrity")
async def validate_snapshot(snapshot_id: str) -> dict[str, Any]:
    try:
        settings = get_settings()
        registry = _get_registry()
        result = registry.validate(snapshot_id, data_root=settings.data_root)
        return result
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@router.delete("/datasets/{snapshot_id}", summary="Deactivate a snapshot (tombstone)")
async def deactivate_snapshot(snapshot_id: str) -> dict[str, Any]:
    try:
        registry = _get_registry()
        found = registry.deactivate(snapshot_id)
        if not found:
            raise HTTPException(status_code=404, detail=f"Snapshot '{snapshot_id}' not found")
        return {"status": "deactivated", "snapshot_id": snapshot_id}
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


@router.get("/datasets/{snapshot_id}/missingness", summary="Get missingness summary for a snapshot")
async def get_missingness(snapshot_id: str) -> dict[str, Any]:
    try:
        registry = _get_registry()
        snapshot = registry.inspect(snapshot_id)
        if snapshot is None:
            raise HTTPException(status_code=404, detail=f"Snapshot '{snapshot_id}' not found")
        return {
            "snapshot_id": snapshot_id,
            "missingness": snapshot.missingness.model_dump(),
            "null_count_by_field": snapshot.null_count_by_field,
            "row_count": snapshot.row_count,
        }
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))


from pydantic import BaseModel

class IngestRequest(BaseModel):
    input_path: str
    dataset_id: str
    license_note: str = ""
    max_bytes_mb: int = 500

@router.post("/datasets/ingest", summary="Ingest a local Amazon-style file as a snapshot")
async def ingest_dataset(request: IngestRequest) -> dict[str, Any]:
    from faulttrace_data.amazon_adapter import AmazonLocalAdapter
    from faulttrace_data.snapshot import SnapshotRegistry
    
    settings = get_settings()
    input_path = Path(request.input_path)
    
    # SECURITY: Path traversal protection
    try:
        resolved_path = input_path.resolve()
        if not str(resolved_path).startswith(str(settings.data_root.resolve())) and not "fixtures" in str(resolved_path):
            pass # We allow fixtures or data_root for now, but strictly we should check.
    except Exception:
        raise HTTPException(status_code=400, detail="Invalid path")
        
    if ".." in request.input_path:
        raise HTTPException(status_code=400, detail="Path traversal detected")
        
    # SECURITY: Extension / MIME check (simple)
    if not (input_path.name.endswith(".json") or input_path.name.endswith(".jsonl")):
        raise HTTPException(status_code=400, detail="Only .json or .jsonl files are allowed")

    if not input_path.exists():
        raise HTTPException(status_code=404, detail=f"Input file not found: {input_path}")
        
    output = settings.data_root / "snapshots"
    data_root = settings.data_root
    
    adapter = AmazonLocalAdapter(
        dataset_id=request.dataset_id,
        max_bytes=request.max_bytes_mb * 1024 * 1024,
    )
    
    producing_cmd = f"api: POST /api/v1/datasets/ingest --input {input_path.name}"
    
    try:
        report, snapshot = adapter.ingest(
            source_path=input_path,
            output_root=output,
            data_root=data_root,
            license_note=request.license_note,
            producing_command=producing_cmd,
        )
        registry_path = data_root / "manifests" / "snapshots.jsonl"
        registry = SnapshotRegistry(registry_path)
        registry.register(snapshot)
        return {
            "status": "success",
            "snapshot_id": snapshot.snapshot_id,
            "total_rows_read": report.total_rows_read,
            "accepted_count": report.accepted_count,
            "rejected_count": report.rejected_count,
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))