# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FaultTrace-RAG API",
    "description": "Counterfactual Fault Localization for Corpus-Level LLM Analytics Pipelines. Provides deterministic corpus data, procedural queries, dual gold evaluation, and traced pipeline execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/artifacts/{artifact_id}/metadata": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Get artifact metadata",
        "description": "Get metadata for a run artifact by run_id.",
        "operationId": "get_artifact_metadata_api_v1_artifacts__artifact_id__metadata_get",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "List all ingested dataset snapshots",
        "operationId": "list_datasets_api_v1_datasets_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by dataset ID",
              "title": "Dataset Id"
            },
            "description": "Filter by dataset ID"
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return active snapshots",
              "default": true,
              "title": "Active Only"
            },
            "description": "Only return active snapshots"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Datasets Api V1 Datasets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{snapshot_id}": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get a specific snapshot by ID",
        "operationId": "get_snapshot_api_v1_datasets__snapshot_id__get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Snapshot Api V1 Datasets  Snapshot Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Datasets"
        ],
        "summary": "Deactivate a snapshot (tombstone)",
        "operationId": "deactivate_snapshot_api_v1_datasets__snapshot_id__delete",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Deactivate Snapshot Api V1 Datasets  Snapshot Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{snapshot_id}/validate": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Validate a snapshot's integrity",
        "operationId": "validate_snapshot_api_v1_datasets__snapshot_id__validate_get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate Snapshot Api V1 Datasets  Snapshot Id  Validate Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{snapshot_id}/missingness": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get missingness summary for a snapshot",
        "operationId": "get_missingness_api_v1_datasets__snapshot_id__missingness_get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Missingness Api V1 Datasets  Snapshot Id  Missingness Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/ingest": {
      "post": {
        "tags": [
          "Datasets"
        ],
        "summary": "Ingest a local Amazon-style file as a snapshot",
        "operationId": "ingest_dataset_api_v1_datasets_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ingest Dataset Api V1 Datasets Ingest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestRequest": {
        "properties": {
          "input_path": {
            "type": "string",
            "title": "Input Path"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "license_note": {
            "type": "string",
            "title": "License Note",
            "default": ""
          },
          "max_bytes_mb": {
            "type": "integer",
            "title": "Max Bytes Mb",
            "default": 500
          }
        },
        "type": "object",
        "required": [
          "input_path",
          "dataset_id"
        ],
        "title": "IngestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}